all the product / product type stuff works.  next step is to fix the editing functionality, specifcially the "click here to edit this form" link.  its probably broken for all edit pages

+++ inventory/src/main/java/com/techvvs/inventory/model/ProductCompositeID.groovy
@@ -1,8 +1,8 @@
 package com.techvvs.inventory.model;
 
 public class ProductCompositeID implements Serializable {
-    private Long product_id;
-    private Long product_type_id;
+    private Integer product_id;
+    private Integer product_type_id;
 
     // Getter and Setter
 }
